게시판 db 접근하는 함수들 board_db.js로 분리
